+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Morphometrics tool using obi-one.
 - Custom brain region resolving.
 - Turn `neuroagent` into an MCP client.
+- Tools for (most) GET endpoints of entitycore
 
 ### Fixed
 - Handle breaking change in entitycore.